refactor: analyser --> scopes; reflection --> reflections
