+++ src/Reflections/RelationClassReflection.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace NunoMaduro\Larastan\Reflection;
+namespace NunoMaduro\Larastan\Reflections;
 
 use PHPStan\Reflection\ClassReflection;
